MetadataTest : Move C++ test into GafferTestModule
This is just an implementation detail of the tests exposed publicly by the GafferTest.MetadataTest Python class. So we also move the bindings for the test into a private submodule, cleaning up the GafferTest namespace some more. This seems like a better way of organising the C++ tests, and should probably become out standard approach.

Also renamed the test to make it more specific. More threading tests are coming...
