+++ src/GafferTestModule/MetadataTest.h
@@ -34,16 +34,14 @@
 //
 //////////////////////////////////////////////////////////////////////////
 
-#ifndef GAFFERTEST_METADATATEST_H
-#define GAFFERTEST_METADATATEST_H
+#ifndef GAFFERTESTMODULE_METADATATEST_H
+#define GAFFERTESTMODULE_METADATATEST_H
 
-#include "GafferTest/Export.h"
-
-namespace GafferTest
+namespace GafferTestModule
 {
 
-GAFFERTEST_API void testMetadataThreading();
+void testConcurrentAccessToDifferentInstances();
 
-} // namespace GafferTest
+} // namespace GafferTestModule
 
-#endif // GAFFERTEST_METADATATEST_H
+#endif // GAFFERTESTMODULE_METADATATEST_H
